design socket

+++ DarwinCore/IPSum.m
@@ -80,8 +80,8 @@ void print_free_memory ()
     /* Stats in bytes */
     natural_t mem_used = (vm_stat.active_count +
                           vm_stat.inactive_count +
-                          vm_stat.wire_count) * pagesize;
-    natural_t mem_free = vm_stat.free_count * pagesize;
+                          vm_stat.wire_count) * (unsigned int)pagesize;
+    natural_t mem_free = vm_stat.free_count * (unsigned int)pagesize;
     natural_t mem_total = mem_used + mem_free;
     NSLog(@"used: %u free: %u total: %u", mem_used, mem_free, mem_total);
 }
